update babel / webpack / jest for react testing. simple react snapshot test passing jest
